+++ common/config.php
@@ -10,6 +10,7 @@
  * 
  * @package 299Ko https://github.com/299Ko/299ko
  */
+ 
 define('VERSION', '1.3.0');
 define('COMMON', ROOT . 'common/');
 define('DATA', ROOT . 'data/');