+++ common/core.class.php
@@ -22,6 +22,11 @@ class core {
     private $js;
     private $css;
     private $locale;
+    
+    /**
+     * Metas are used by plugins to display metas property or other in <head> HTML
+     */
+    private $metas = [];
 
     /**
      * Logger is a resource file
@@ -33,8 +38,10 @@ class core {
 
     public function __construct() {
         
-        $this->createLogger();
-        
+        if (!is_dir(DATA))
+            mkdir(DATA);
+		$this->createLogger();
+
         // Timezone
         date_default_timezone_set(date_default_timezone_get());
         // Construction du tableau de configuration
@@ -147,6 +154,14 @@ class core {
 
     public function getCss() {
         return $this->css;
+    }
+    
+    public function addMeta(string $meta) {
+        $this->metas[] = $meta;
+    }
+    
+    public function getMetas() {
+        return $this->metas;
     }
 
     ## Détermine si 299ko est installé